rename LocalProfile Package
rename localprofile to local_profile to match naming convention

fix #10436
