+++ src/main/java/tech/jhipster/lite/generator/server/springboot/locale_profile/application/LocalProfileApplicationService.java
@@ -1,7 +1,7 @@
-package tech.jhipster.lite.generator.server.springboot.localeprofile.application;
+package tech.jhipster.lite.generator.server.springboot.locale_profile.application;
 
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.server.springboot.localeprofile.domain.LocalProfileModuleFactory;
+import tech.jhipster.lite.generator.server.springboot.locale_profile.domain.LocalProfileModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
